Changes communication profile to MQTT
Orginally was reading a webpage that's continually updated via a
cronjob. Now is updated by a message sent from the arduino client,
which submits the voltage of the display. This voltage with the
weather data is submitted to graphite on the server side. The server
daemon also submits the weather data to sensor to display.